requirements.txt
# Data Import
playwright>=1.45.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Data Processing
pandas>=2.0.0
pydantic>=2.0.0
numpy>=1.24.0

# PII Detection
presidio-analyzer>=2.2.0
thefuzz>=0.19.0

# Embeddings & Clustering
sentence-transformers>=2.2.0
hdbscan>=0.8.33
scikit-learn>=1.3.0

# LLM Integration
google-generativeai>=0.8.0
google-api-python-client>=2.137.0
google-auth>=2.26.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
python-dateutil>=2.8.2
pytest>=7.4.0

